Moved InstrumentedHandler to a method on the server to save passing state around. Addressed other review feedback.
