+++ internal/handlers.go
@@ -2,30 +2,10 @@ package fleetlock
 
 import (
 	"net/http"
-	"strconv"
-
-	"github.com/prometheus/client_golang/prometheus"
 )
-
-type responseWriter struct {
-	http.ResponseWriter
-	statusCode int
-}
-
-// Makes a responseWriter
-func newResponseWriter(w http.ResponseWriter) *responseWriter {
-	return &responseWriter{w, http.StatusOK}
-}
-
-// Writes the header code into the response and makes it available for logging
-func (rw *responseWriter) WriteHeader(code int) {
-	rw.statusCode = code
-	rw.ResponseWriter.WriteHeader(code)
-}
 
 const (
 	fleetLockHeaderKey = "fleet-lock-protocol"
-	millisecondsInSec  = 1000
 )
 
 // POSTHandler returns a handler that requires the POST method.
@@ -48,27 +28,6 @@ func HeaderHandler(key, value string, next http.Handler) http.Handler {
 			return
 		}
 		next.ServeHTTP(w, req)
-	}
-	return http.HandlerFunc(fn)
-}
-
-// InstrumentedHandler returns a handler that instruments http requests
-func InstrumentedHandler(totalRequests, responseStatus *prometheus.CounterVec, httpDuration *prometheus.HistogramVec, next http.Handler) http.Handler {
-	fn := func(w http.ResponseWriter, req *http.Request) {
-		path := req.URL.Path
-		timer := prometheus.NewTimer(prometheus.ObserverFunc(func(v float64) {
-			ms := v * millisecondsInSec // make microseconds
-			httpDuration.WithLabelValues(path).Observe(ms)
-		}))
-		defer timer.ObserveDuration()
-		rw := newResponseWriter(w)
-		next.ServeHTTP(rw, req)
-
-		responseStatus.With(prometheus.Labels{
-			"path":   path,
-			"status": strconv.Itoa(rw.statusCode),
-		}).Inc()
-		totalRequests.WithLabelValues(path).Inc()
 	}
 	return http.HandlerFunc(fn)
 }
